Add tests to exhibit RegExp constructor behaviour
Exhibit the behavior of `new RegExp(str, flags)`. The main reason is to
describe the current state of the parser before making changes to it.

The tests show which combinations of flags make parsing raise and
`toString` of the resulting `RegExp` which makes it clearer how some
inputs are interpreted.