Allow empty branches for unions and intersections
Previously `a|` would fail (expecting more after the `|`) and `|a`
parses equivalently to `"|a"`. Instead we parse both as a union of `a`
and `()`. We make a similar change to `&`. A lone `|` parses as `""|"""`
and likewise for a lone `&`. A side-effect is that we no-longer allow
parsing `)` as a RegExp.

We change the grammar description as follows:
1. We move the empty case from _regexp_ to go between _interexp_ and
   _concatexp_. The new symbol gets the name _optionalexp_.
2. We remove the `()` production rule from _simpleexp_ as it is now
   redundant: a _unionexp_ may be empty.