Refuse to parse repeat operators like '*' as chars
If part of a regexp starts with `*`, `+`, or `?` (e.g. at the start or
after a `|` or `(`), we would previously parse this as a literal (i.e.
matching the character itself). We change that to instead result in an
IllegalArgumentException. We also defensively reject other reserved
characters like `(` or `"`, though there ought not be any code paths to
`parseCharExp` where such chars are next.

We also change the documented grammar to make it more clear which
characters are reserved, splitting out character-in-charclass into a
separate symbol. The grammar is still a bit ambiguous about char class
parsing.

Finally, tests are updated to match the new behaviour.